Drop disqus, update about and upload draft articles
I'm going to drop disqus for comments for now, as I have concerns on the
security of that service towards its users. I will be using Twitter for
gathering public feedback, or e-mails otherwise.

The about page is updated with a blurb about self-employment.

Finally, a few draft articles are uploaded which I intend to fine-tune
further and release later.